v0.0.1.1
Major update: code now structured as an R package

The main folder now includes a GenEst package subfolder, the top level of which has a script of development code and the compiled current version of the package, as well as the directory for the package in development.

That directory follows standard formatting for R packages

Format now following Wickham's manual for developing R packages

Function, object, and argument names have all been standardized in format and convention

The function code has now been broken into separate scripts: search efficiency, carcass persistence, detection probability, fatality, the gui, and utility functions.

The command line example has been converted to a vignette that allows for a simple tutorial for the user at the command line.

The app has been included as installed files called via a runGenEst function

In addition, an error in the SEmod functions associated with the calcualtion of AICc values was fixed.