tidy and edit SE functions
clean pkm function

note edit to calculation of the k cell means. previous version was using values from the solved hessian rather than the point MLE values
in addtion, there are a bunch of naming changes here. the goal with the naming and the general formatting is for a fluid, easy-to-read code base that will promote usage, development, and bug fixing. names of objects are therefore designed to be intuitive, readable, and as simple as possible.
this will necessarily mean that other functions will also need to be edited to work properly, and that's coming down the pike next.