cbinom package (source..primarily for Mac & Linux)
can be used in Windows but requires that your  Rtools C++ compiling tool chain in R is properly set up on your computer. If it's not or you aren't sure, you can use the pre-compiled Windows binary version (.zip). Alternatively, you can first install Rtools and hope it works for you (it may work perfectly out of the box...or require some extra effort fixing some path parameters and whatnot). And then install the .tar.gz version of the package.