experimenting
everything works fine on my desktop with a clean R session 
everything passes the travis tests fine until I ask to merge...then fails w/in some vignette graph. Which one? 
Commented out much good code in vignettes (to speed the testing)
commented out part of the offending code (not sure if the code itself is offensive or whether Travis is just taking it the wrong way) in the spirit of testing.