Delete estgGenericCell.Rd
name of function changed to "calcg" because there is no random component in the calculations, i.e., for given parameters, calcg calculates g. The other "gGeneric" functions instead simulate from pkmod and cpmod to estimate g, so the estg prefix is retained. 

Also, calcg has utility beyond the app tabiya (estgGeneric, estgGenericSize, estgGenericCell), so the more general "calcg" is used in place of the idiomatic "estgGenericCell".