v 0.2.7 GUI edits
reformatted the generic g summary table akin to the M summary table

fixed the buggy situation with the selection of models. now there is no default selection and there is proper handling of no selection being made yet. further, when a model is re-run, the selection is functionally wiped out, and so needs to be made again (eliminating a source of confusion)