v0.2.10 removal of size class text and dropdowns when analysis done without size class
the SE CP and g analyses all have size class components (potentially) to the output, and previously when a size class was not used, the size class options were still presented but with the placeholder "all". now those components are removed. this includes the drop downs as well as the text on the output tabs

in addition, i've added some explanatory text to the model selection tabs to guide the user and removed the superfluous text