fully developed server function
now we have a GenEstServer function that is expressed within the server.R script
to help tidy the code, we have a set of functions that provide the reaction for each event that is observed (reaction, reactionMessageRun, reactionMessageDone, and eventReaction), all of which are documented within the server function's documentation

to bring everything in alignment, a few extra update_ functions had to be made, but they (as well as the code within eventReaction) will be substantially tightened up shortly. i just needed to start at the top and work downward from there.