Dan (#618)
* Update code.json

* in progress

* error msg for plot.pkm/Set when k missing (GUI) (#607)

* Add files via upload

* Add files via upload

* Add files via upload

* Add files via upload

* Add files via upload

* license to CC0 + .RbuildIgnore LICENSE.md

* Update .Rbuildignore

A few files that are necessary for USGS but are frown-inducing at CRAN (I don't remember if it is "Warning" or "Note"--either way, it's not good) are now included in the .Rbuildignore file.

* in progress more

stable stopping point

* in progress some more

server script is tidied up now and set as functions

* fully developed server function

now we have a GenEstServer function that is expressed within the server.R script
to help tidy the code, we have a set of functions that provide the reaction for each event that is observed (reaction, reactionMessageRun, reactionMessageDone, and eventReaction), all of which are documented within the server function's documentation

to bring everything in alignment, a few extra update_ functions had to be made, but they (as well as the code within eventReaction) will be substantially tightened up shortly. i just needed to start at the top and work downward from there.

* condense all update_input functions

into a single function with an eventName argument

* documentation edits

args etc

* in prog

* condensed output update functions

now a single update_output function for the app, akin to the update_input function that works on the eventName and goes from there

* rv updating now like output and input updating

substantial reduction in number of functions for rv updates, akin to the other two sets of functions
updated vignette to include server details

* minor edits

* Update DESCRIPTION

* fixing output suspension

output objects (figures and tables) had been getting not suspended when hidden, which was causing odd lacks of updating. this is fixed here, resolving the proximate issue.  there might be knock on effects for conditionals, but they can be addressed separetly

* fix the conditionals

the corresponding conditionals for the model objects are now set to work on the moddone output text, not the table or figure itself

* handling the resetting of variables properly

there were some order of operations issues for the clearing of rv and output
it's now set up that basically everything relevant is cleared at the beginning of a run, and then repopulated if success occurs, and then any handling of options (i.e, suspending when hidden) occurs after the run
that is, except for splits, which are being addressed separatelyy

* handle messages for splits

splits messaging wasn't working due to an issue in the conditional within the message function

* code tidying

removing vestigial functions: disclaimers (has been replaced by disclaimerUSGS and disclaimerWEST)
removing all references and uses of kfill in the GUI. in the situations where there was remaining utility (i.e. noting when k wasnt set in the SE model and someone is interested in M or g), kFillNeed and setkFillNeed are now kNeed and setkNeed
updated vignette

* Update app_content.R

* Update DESCRIPTION

* Dan (#613)

* permanent links, more robust error-checking in ppersist

* Update carcass_persistence_functions.R

* Update DESCRIPTION

* Update README.md

* Update carcass_persistence_functions.R

* Update DESCRIPTION

* fix parsing estM args for better GUI compatibility

* correct typo

* links to published TM docs; minor bug fixes

* updated links; minor bug fixes

* Update README.md

* updated error checks for dates and units

* fix typo

* link to doi rather than pdf

* json update

* sample size in s '+' models

* v1.1.1